requirements.txt
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.7
diskcache==5.6.3
idna==3.10
Jinja2==3.1.6
llama_cpp_python==0.3.16
MarkupSafe==3.0.2
numpy==2.3.2
packaging==25.0
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pip==23.2.1
pycparser==2.22
PyMuPDF==1.26.4
pypdfium2==4.30.0
pytesseract==0.3.13
requests==2.32.5
setuptools==68.2.0
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.41.2
